requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Models & validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# MCP SDK
mcp>=1.0.0

# Google Gemini
google-genai>=1.0.0

# PDF processing
PyPDF2>=3.0.0
httpx>=0.27.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
httpx>=0.27.0

# Linting
ruff>=0.6.0
